Update main.lua to properly fix time check
This change includes implementing a ticket required to access. on top of that the server was going off of SERVER time not INGAME TIME. While this is more secure it can cause issues with players if the black market  is not available when the brunt of population is online. With this fix it runs off of INGAME TIME and still checks for the job and ticket so that it cant be exploited.